添加信息修改

+++ db/models/ShopsInfos.js
@@ -5,8 +5,8 @@ const shopsinfo = db.defineModel('shopsinfo', {
   name: db.STRING, // 店铺名
   details: db.STRING, // 店铺详细地址
   area: db.STRING, // 店铺所在区域
-  longitude: db.DECIMAL, // 店铺经度
-  latitude: db.DECIMAL, // 店铺纬度
+  longitude: db.DOUBLE, // 店铺经度
+  latitude: db.DOUBLE, // 店铺纬度
   tele: db.BIGINT, // 店铺电话
   rate: db.INTEGER // 店铺评分
 })
